+++ Changelog.md
@@ -12,6 +12,7 @@
 - Fix remote print initialize at startup
 - Fix http status proxy checkbox
 - Fix being unable to start new remote print after stopping one
+- Allow stopping prints through remote print V1
 
 ## v0.9.1 &mdash; August 28th, 2026